% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi.R
\name{nmi}
\alias{nmi}
\alias{NMI}
\title{Normalized Mutual Information (NMI)}
\source{
\itemize{
\item Cover, T. M. & Thomas, J. A. (1991) \emph{Elements of Information Theory}. New York, NY: John Wiley & Sons.
\item \href{https://en.wikipedia.org/wiki/Mutual_information#Normalized_variants}{Normalized variants of the mutual information}
}
}
\usage{
nmi(y, yhat)
}
\arguments{
\item{y}{actual class labels or predicted cluster labels to compare against}

\item{yhat}{predicted cluster labels (not necessarily the same absolute
values; labels may be permuted)}
}
\description{
NMI is a normalization of the Mutual Information (MI) score to
scale the results between 0 (no mutual information) and 1
(perfect correlation). Can also be used to assess agreement between two
clusterings.
}
\examples{
\dontrun{
set.seed(0)
clusters <- kmeans(iris[, 1:4], centers = 3)
y <- as.numeric(iris$Species)
yhat <- clusters$cluster
nmi(y, yhat) # 0.76

data("wine")
set.seed(0)
clusters <- kmeans(wine[, -1], centers = 3)
y <- wine$class
yhat <- clusters$cluster
nmi(y, yhat) # 0.43
}
}
