% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine Data Set}
\format{A \code{data.frame} with 178 rows and 13 variables:
\describe{
\item{class}{\href{https://en.wikipedia.org/wiki/Cultivar}{Cultivar}}
\item{alcohol}{\href{https://en.wikipedia.org/wiki/Alcohol}{Alcohol}}
\item{malic_acid}{\href{https://en.wikipedia.org/wiki/Malic_acid}{Malic acid} contributes to the pleasantly sour taste of fruits, and is used as a food additive}
\item{ash}{Non-aqueous, non-gaseous residues that remain after something is burned}
\item{ash_alcalinity}{Capacity to resist changes in pH}
\item{magnesium}{\href{https://en.wikipedia.org/wiki/Magnesium}{Magnesium}}
\item{total_phenols}{The \href{https://en.wikipedia.org/wiki/Phenolic_content_in_wine}{phenolic content in wine} refers to the phenolic compounds -- natural phenol and polyphenols -- in wine}
\item{flavanoids}{\href{https://en.wikipedia.org/wiki/Flavonoid}{Flavonoids} (from the Latin word \emph{flavus} meaning yellow, their color in nature) are a class of plant and fungus secondary metabolites}
\item{nonflavanoid_phenols}{\url{}}
\item{proanthocyanins}{\href{https://en.wikipedia.org/wiki/Condensed_tannin}{Proanthocyanidins} are polymers formed by the condensation of flavans; they do not contain sugar residues}
\item{color_intensity}{\href{https://en.wikipedia.org/wiki/Colorfulness#Saturation}{Saturation} of a color is determined by a combination of light intensity and how much it is distributed across the spectrum of different wavelengths}
\item{hue}{\href{https://en.wikipedia.org/wiki/Hue}{Hue} is one of main properties of color}
\item{protein_concentration}{OD280/OD315 of diluted wines}
\item{proline}{\href{https://en.wikipedia.org/wiki/Proline}{Proline} is a proteinogenic amino acid that is used in the biosynthesis of proteins}
}}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/wine}{UCI Machine Learning Repository}
}
\usage{
wine
}
\description{
Using chemical analysis determine the origin of wines.
}
\references{
\itemize{
\item Forina, M. et al, PARVUS - An Extendible Package for Data Exploration,
Classification and Correlation. Institute of Pharmaceutical and Food
Analysis and Technologies, Via Brigata Salerno, 16147 Genoa, Italy.
\item Stefan Aeberhard
}
}
\keyword{datasets}
